[features] move sift describer in openMVG/features
