+++ src/openMVG_Samples/robust_essential/robust_essential.cpp
@@ -12,7 +12,7 @@
 #include "openMVG/matching/regions_matcher.hpp"
 #include "openMVG/multiview/triangulation.hpp"
 
-#include "nonFree/sift/SIFT_describer.hpp"
+#include "openMVG/features/sift/SIFT_describer.hpp"
 
 #include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
 #include "third_party/vectorGraphics/svgDrawer.hpp"
