+++ src/openMVG_Samples/robust_essential_ba/robust_essential_ba.cpp
@@ -17,7 +17,7 @@
 
 #include "openMVG/matching/regions_matcher.hpp"
 
-#include "nonFree/sift/SIFT_describer.hpp"
+#include "openMVG/features/sift/SIFT_describer.hpp"
 
 #include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
 #include "third_party/vectorGraphics/svgDrawer.hpp"
