+++ src/openMVG_Samples/robust_essential_spherical/robust_essential_spherical.cpp
@@ -14,7 +14,7 @@
 #include "openMVG/multiview/conditioning.hpp"
 #include "openMVG/robust_estimation/robust_estimator_ACRansacKernelAdaptator.hpp"
 
-#include "nonFree/sift/SIFT_describer.hpp"
+#include "openMVG/features/sift/SIFT_describer.hpp"
 #include "openMVG_Samples/robust_essential_spherical/spherical_cam.hpp"
 
 #include "software/SfM/SfMPlyHelper.hpp"
