+++ src/openMVG_Samples/robust_fundamental/robust_fundamental.cpp
@@ -13,7 +13,7 @@
 #include "openMVG/robust_estimation/robust_estimator_ACRansac.hpp"
 #include "openMVG/robust_estimation/robust_estimator_ACRansacKernelAdaptator.hpp"
 
-#include "nonFree/sift/SIFT_describer.hpp"
+#include "openMVG/features/sift/SIFT_describer.hpp"
 #include "third_party/cmdLine/cmdLine.h"
 
 #include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
