+++ src/openMVG_Samples/robust_homography_guided/robust_homography_guided.cpp
@@ -15,7 +15,7 @@
 
 #include "openMVG/robust_estimation/guided_matching.hpp"
 
-#include "nonFree/sift/SIFT_describer.hpp"
+#include "openMVG/features/sift/SIFT_describer.hpp"
 
 #include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
 #include "third_party/vectorGraphics/svgDrawer.hpp"
